Move alternative libraries tests to a separate file and tweak the README regarding supported libraries
