+++ deno.lock
@@ -505,16 +505,19 @@
   "npm": {
     "specifiers": {
       "@deepkit/type@1.0.1-alpha.97": "@deepkit/type@1.0.1-alpha.97_@deepkit+core@1.0.1-alpha.97",
+      "@effect/schema": "@effect/schema@0.34.0_@effect+data@0.18.3_@effect+io@0.39.1__@effect+data@0.18.3",
       "@sinclair/typebox": "@sinclair/typebox@0.31.0",
       "@sinclair/typebox@0.31.0": "@sinclair/typebox@0.31.0",
       "ajv": "ajv@8.12.0",
       "ajv@8.12.0": "ajv@8.12.0",
+      "arktype": "arktype@0.6.0",
       "arktype@1.0.19-alpha": "arktype@1.0.19-alpha",
       "io-ts": "io-ts@2.2.20_fp-ts@2.16.1",
       "io-ts@2.2.20": "io-ts@2.2.20_fp-ts@2.16.1",
       "joi": "joi@17.9.2",
       "joi@17.9.2": "joi@17.9.2",
       "ow@0.28.2": "ow@0.28.2",
+      "runtypes": "runtypes@6.7.0",
       "runtypes@6.7.0": "runtypes@6.7.0",
       "superstruct": "superstruct@1.0.3",
       "superstruct@1.0.3": "superstruct@1.0.3",
@@ -543,6 +546,24 @@
           "@types/uuid": "@types/uuid@8.3.4",
           "buffer": "buffer@5.7.1",
           "uuid": "uuid@8.3.2"
+        }
+      },
+      "@effect/data@0.18.3": {
+        "integrity": "sha512-xNG4CwxE758EK0v81EIPnKv9QeU/x9st1Doab/Fjf7sb9V3n55C85aBiKChpKgGD8pKCPHcjGydLsRHVUvzpwg==",
+        "dependencies": {}
+      },
+      "@effect/io@0.39.1_@effect+data@0.18.3": {
+        "integrity": "sha512-ISBX1EYsWdVBMOV3ZmEVtbMW8FyVwmK6K47rx8ztOjMBKS5tfhB4W8iRjyycObvvlolfpIcZuAAPrCW5Tty+4Q==",
+        "dependencies": {
+          "@effect/data": "@effect/data@0.18.3"
+        }
+      },
+      "@effect/schema@0.34.0_@effect+data@0.18.3_@effect+io@0.39.1__@effect+data@0.18.3": {
+        "integrity": "sha512-7BdriGSWYj+f9jVqIq6PyfLoPztV05RW1Sj+mdKX3gAwqqIM7MBwkM0kTaLYKAFbGXTE9AK83uG2SSfzMigqTA==",
+        "dependencies": {
+          "@effect/data": "@effect/data@0.18.3",
+          "@effect/io": "@effect/io@0.39.1_@effect+data@0.18.3",
+          "fast-check": "fast-check@3.12.1"
         }
       },
       "@hapi/hoek@9.3.0": {
@@ -605,6 +626,10 @@
         "dependencies": {
           "color-convert": "color-convert@2.0.1"
         }
+      },
+      "arktype@0.6.0": {
+        "integrity": "sha512-SBxj1wD3GIG9zzjaseksVs2zwKXb/fjyCHr9grpC0oaror1NQWylkzr6HtMkYIusxGuwB33OfhYZ78kCFDCZbw==",
+        "dependencies": {}
       },
       "arktype@1.0.19-alpha": {
         "integrity": "sha512-tcmWSZupIjjG/hBYAlL9JzheY3WP3upCPlD7UCbY65AbzZegLZXrbCM4+TI03t2rKRwT2TuSioH27Bdqst757Q==",
@@ -734,6 +759,12 @@
           "chardet": "chardet@0.7.0",
           "iconv-lite": "iconv-lite@0.4.24",
           "tmp": "tmp@0.0.33"
+        }
+      },
+      "fast-check@3.12.1": {
+        "integrity": "sha512-aRN6WrO+q3TZV8CjXnpebx9bTHrpYIOVi6v3ttxciZiWqS739yy8b1oVx+qNEyV1b8RdVjlp/+miTY6yAp90HA==",
+        "dependencies": {
+          "pure-rand": "pure-rand@6.0.2"
         }
       },
       "fast-deep-equal@3.1.3": {
@@ -891,6 +922,10 @@
       },
       "punycode@2.3.0": {
         "integrity": "sha512-rRV+zQD8tVFys26lAGR9WUuS4iUAngJScM+ZRSKtvl5tKeZ2t5bvdNFdNHBW9FWR4guGHlgmsZ1G7BSm2wTbuA==",
+        "dependencies": {}
+      },
+      "pure-rand@6.0.2": {
+        "integrity": "sha512-6Yg0ekpKICSjPswYOuC5sku/TSWaRYlA0qsXqJgM/d/4pLPHPuTxK7Nbf7jFKzAeedUhR8C7K9Uv63FBsSo8xQ==",
         "dependencies": {}
       },
       "randexp@0.5.3": {
