+++ src/constructor.test.ts
@@ -5,14 +5,6 @@ import {
   it,
 } from './test-prelude.ts'
 import { z } from 'https://deno.land/x/zod@v3.21.4/mod.ts'
-import * as v from 'npm:valibot'
-import * as y from 'npm:yup'
-import * as t from 'npm:io-ts'
-import typia from 'npm:typia'
-import joi from 'npm:joi'
-import { Type as TypeBox } from 'npm:@sinclair/typebox'
-import ajv from 'npm:ajv' // This import is used indirectly by the runtime
-import * as s from 'npm:superstruct'
 
 import { makeDomainFunction } from './constructor.ts'
 import {
@@ -36,139 +28,6 @@ describe('makeDomainFunction', () => {
       inputErrors: [],
       environmentErrors: [],
     })
-  })
-
-  describe('when using different libraries to create schemas', () => {
-    it.ignore(
-      'can use a typia schema - test type only, typiq requires a runtime transformer to run validators',
-      async () => {
-        const schema = typia.createAssert<{ id: number }>()
-        const handler = makeDomainFunction(schema)(({ id }) => id)
-        type _R = Expect<Equal<typeof handler, DomainFunction<number>>>
-
-        assertEquals(await handler({ id: 1 }), {
-          success: true,
-          data: 1,
-          errors: [],
-          inputErrors: [],
-          environmentErrors: [],
-        })
-      },
-    )
-
-    it('can use an superstruct schema', async () => {
-      const schema = s.object({ id: s.number() })
-
-      const handler = makeDomainFunction(schema)(({ id }) => id)
-      type _R = Expect<Equal<typeof handler, DomainFunction<number>>>
-
-      assertEquals(await handler({ id: 1 }), {
-        success: true,
-        data: 1,
-        errors: [],
-        inputErrors: [],
-        environmentErrors: [],
-      })
-    })
-    it('can use an joi schema (no type inference)', async () => {
-      const schema = joi.object({ id: joi.number() })
-
-      const handler = makeDomainFunction(schema)(({ id }: any) => id)
-
-      assertEquals(await handler({ id: 1 }), {
-        success: true,
-        data: 1,
-        errors: [],
-        inputErrors: [],
-        environmentErrors: [],
-      })
-    })
-
-    it('can use an ajv schema (no type inference)', async () => {
-      const schema = {
-        type: 'object',
-        properties: {
-          id: { type: 'integer' },
-        },
-        required: ['id'],
-        additionalProperties: false,
-      }
-
-      const handler = makeDomainFunction(schema)(({ id }: any) => id)
-
-      assertEquals(await handler({ id: 1 }), {
-        success: true,
-        data: 1,
-        errors: [],
-        inputErrors: [],
-        environmentErrors: [],
-      })
-    })
-
-    it('can use a typebox schema', async () => {
-      const schema = TypeBox.Object({ id: TypeBox.Integer() })
-
-      const handler = makeDomainFunction(schema)(({ id }) => id)
-      type _R = Expect<Equal<typeof handler, DomainFunction<number>>>
-
-      assertEquals(await handler({ id: 1 }), {
-        success: true,
-        data: 1,
-        errors: [],
-        inputErrors: [],
-        environmentErrors: [],
-      })
-    })
-
-    it('can use a valibot schema', async () => {
-      const schema = v.object({ id: v.number() })
-
-      const handler = makeDomainFunction(schema)(({ id }) => id)
-      type _R = Expect<Equal<typeof handler, DomainFunction<number>>>
-
-      assertEquals(await handler({ id: 1 }), {
-        success: true,
-        data: 1,
-        errors: [],
-        inputErrors: [],
-        environmentErrors: [],
-      })
-    })
-
-    it('can use a yup schema', async () => {
-      const schema = y.object({ id: y.number() })
-
-      const handler = makeDomainFunction(schema)(({ id }) => id)
-      type _R = Expect<
-        Equal<typeof handler, DomainFunction<number | undefined>>
-      >
-
-      assertEquals(await handler({ id: 1 }), {
-        success: true,
-        data: 1,
-        errors: [],
-        inputErrors: [],
-        environmentErrors: [],
-      })
-    })
-
-    it.ignore(
-      'can use a io-ts schema - keep here to test the type, runtime has issues from deno',
-      async () => {
-        const schema = t.type({ id: t.number })
-
-        const handler = makeDomainFunction(schema)(({ id }) => id)
-        type _R = Expect<Equal<typeof handler, DomainFunction<number>>>
-
-        assertEquals(await handler({ id: 1 }), {
-          success: true,
-          data: 1,
-          errors: [],
-          inputErrors: [],
-          environmentErrors: [],
-        })
-      },
-    )
   })
 
   describe('when it has no environment', () => {
